Break out sort and map functions
